Add navigation
Update App.js
Delete custom icon files.
Add images.
Add router.js for tab and stack navigation.
Create Background wrapper with reexport.
Move files and folder in screens.
Add screens CommentsScreen,Home,MapScreen,PostsScreen,ProfileScreen
