+++ src/Components/index.js
@@ -1,0 +1,1 @@
+export { default as Background } from "./Background/Background";
