+++ src/Screens/index.js
@@ -1,2 +1,11 @@
-export { default as LoginScreen } from "./LoginScreen";
-export { default as RegistrationScreen } from "./RegistrationScreen";
+// Auth
+
+export { default as LoginScreen } from "./auth/LoginScreen";
+export { default as RegistrationScreen } from "./auth/RegistrationScreen";
+
+// Main
+
+export { default as ProfileScreen } from "./main/ProfileScreen";
+export { default as CreatePostsScreen } from "./main/CreatePostsScreen";
+export { default as PostsScreen } from "./main/PostsScreen";
+export { default as Home } from "./main/Home";
